+++ xknx/management/procedures/network/nm_individual_address_read.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import logging
 
 from xknx.exceptions import ManagementConnectionError
-from xknx.management.protocols import Broadcaster
+from xknx.management.management import Management
 from xknx.telegram import apci
 from xknx.telegram.address import IndividualAddress
 
@@ -13,21 +13,21 @@ logger = logging.getLogger("xknx.management.procedures")
 
 
 async def nm_individual_address_read(
-    broadcaster: Broadcaster,
+    management: Management,
     timeout: float | None = 3,
     raise_if_multiple: bool = False,
 ) -> list[IndividualAddress]:
     """
     Request individual addresses of all devices that are in programming mode.
 
-    :param broadcaster: broadcaster for sending and receiving broadcast telegrams
+    :param management: connection manager used for sending and receiving broadcast telegrams
     :param timeout: specifies the timeout in seconds, the KNX specification requires a timeout of 3s
     :param raise_if_multiple: if true, ManagementConnectionError is raised when multiple devices are in programming mode
     :returns: list of individual address of devices in programming mode
     """
     addresses = []
-    async with broadcaster.broadcast() as bc_context:
-        await broadcaster.send_broadcast(apci.IndividualAddressRead())
+    async with management.broadcast() as bc_context:
+        await management.send_broadcast(apci.IndividualAddressRead())
         async for result in bc_context.receive(timeout=timeout):
             if isinstance(result.payload, apci.IndividualAddressResponse):
                 addresses.append(result.source_address)