+++ xknx/management/procedures/network/nm_individual_address_serial_number_read.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import logging
 
-from xknx.management.protocols import Broadcaster
+from xknx.management.management import Management
 from xknx.telegram import apci
 from xknx.telegram.address import IndividualAddress
 
@@ -12,13 +12,13 @@ logger = logging.getLogger("xknx.management.procedures")
 
 
 async def nm_individual_address_serial_number_read(
-    broadcaster: Broadcaster,
+    management: Management,
     serial: bytes,
     timeout: float = 3,
 ) -> IndividualAddress | None:
     """Read individual address from device with specified serial number."""
-    async with broadcaster.broadcast() as bc_context:
-        await broadcaster.send_broadcast(
+    async with management.broadcast() as bc_context:
+        await management.send_broadcast(
             payload=apci.IndividualAddressSerialRead(serial=serial)
         )
         async for result in bc_context.receive(timeout=timeout):