+++ xknx/management/procedures/network/nm_individual_address_serial_number_write.py
@@ -5,10 +5,10 @@ from __future__ import annotations
 import logging
 
 from xknx.exceptions import ManagementConnectionError
+from xknx.management.management import Management
 from xknx.management.procedures.network.nm_individual_address_serial_number_read import (
     nm_individual_address_serial_number_read,
 )
-from xknx.management.protocols import Broadcaster
 from xknx.telegram import apci
 from xknx.telegram.address import IndividualAddress, IndividualAddressableType
 
@@ -16,13 +16,13 @@ logger = logging.getLogger("xknx.management.procedures")
 
 
 async def nm_individual_address_serial_number_write(
-    broadcaster: Broadcaster,
+    management: Management,
     serial: bytes,
     individual_address: IndividualAddressableType,
 ) -> None:
     """Write individual address to device with specified serial number."""
     individual_address = IndividualAddress(individual_address)
-    await broadcaster.send_broadcast(
+    await management.send_broadcast(
         payload=apci.IndividualAddressSerialWrite(
             address=individual_address,
             serial=serial,
@@ -34,7 +34,7 @@ async def nm_individual_address_serial_number_write(
         serial,
     )
 
-    address = await nm_individual_address_serial_number_read(broadcaster, serial=serial)
+    address = await nm_individual_address_serial_number_read(management, serial=serial)
 
     if address is None:
         raise ManagementConnectionError(f"No reply received from {serial!r}.")